# backend/swagger/openapi.sample.yaml
openapi: 3.0.3
info:
  title: Lost Person Car Detection API
  version: 1.0.0
  description: Sample OpenAPI contract for the backend
servers:
  - url: http://localhost:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GeoPoint:
      type: object
      required: [coordinates]
      properties:
        type:
          type: string
          example: Point
        coordinates:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
        address:
          type: string
paths:
...
  /api/v1/detections/vehicles:
    post:
      tags: [Car Detection]
      summary: Report a vehicle detection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, location]
              properties:
                description:
                  type: string
                location:
                  $ref: "#/components/schemas/GeoPoint"
      responses:
        "201":
          description: Vehicle detection created
...